requirements.txt
python-docx
mammoth
tkinterweb
pyinstaller
